refactor(kmod): split agnocast_main.c into init, exit
